+++ src/services/ApolloService.js
@@ -1,4 +1,4 @@
-class ApolloService
+export class ApolloService
 {
     /*
     * ApolloService Class
@@ -56,12 +56,12 @@ class ApolloService
     }
 
     // correlates to below
-    GetFakeWormGeneData(){
+    GetFakeWormGeneDataEgl8(){
       // http://localhost:8080/apollo/track/worm/All%20Genes/V:17405..47604.json?includeGenotypes=false&ignoreCache=true
       return [{"strand":1,"children":[{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":31880,"type":"exon","fmax":31901,"seqId":"V"}],"name":"B0348.7","source":"WormBase","id":"Transcript:B0348.7","fmin":31880,"type":"piRNA","fmax":31901,"seqId":"V"}],"name":"21ur-15398","source":"WormBase","id":"http://localhost:8080/apollo/track/worm/All Genes/V/21ur-15398.json","fmin":31880,"type":"gene","fmax":31901,"seqId":"V"},{"strand":1,"children":[{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":39258,"type":"exon","fmax":39364,"seqId":"V"}],"name":"B0348.8","source":"WormBase","id":"Transcript:B0348.8","fmin":39258,"type":"snoRNA","fmax":39364,"seqId":"V"}],"name":"B0348.8","source":"WormBase","id":"http://localhost:8080/apollo/track/worm/All Genes/V/B0348.8.json","fmin":39258,"type":"gene","fmax":39364,"seqId":"V"},{"strand":1,"children":[{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4n","source":"WormBase","id":"CDS:B0348.4n","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4n","source":"WormBase","id":"CDS:B0348.4n","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4n","source":"WormBase","id":"CDS:B0348.4n","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4n","source":"WormBase","id":"CDS:B0348.4n","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4n","source":"WormBase","id":"CDS:B0348.4n","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4n","source":"WormBase","id":"CDS:B0348.4n","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4n","source":"WormBase","id":"CDS:B0348.4n","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4n","source":"WormBase","id":"CDS:B0348.4n","fmin":28381,"type":"CDS","fmax":28699,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4n","source":"WormBase","id":"CDS:B0348.4n","fmin":29335,"type":"CDS","fmax":29515,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4n","source":"WormBase","id":"CDS:B0348.4n","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4n","source":"WormBase","id":"CDS:B0348.4n","fmin":31339,"type":"CDS","fmax":31582,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4n","source":"WormBase","id":"CDS:B0348.4n","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4n","source":"WormBase","id":"CDS:B0348.4n","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4n","source":"WormBase","id":"CDS:B0348.4n","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4n","source":"WormBase","id":"CDS:B0348.4n","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4n","source":"WormBase","id":"CDS:B0348.4n","fmin":38990,"type":"CDS","fmax":39026,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4n","source":"WormBase","id":"CDS:B0348.4n","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4n","source":"WormBase","id":"CDS:B0348.4n","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4n","source":"WormBase","id":"CDS:B0348.4n","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4n","source":"WormBase","id":"CDS:B0348.4n","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4n","source":"WormBase","id":"CDS:B0348.4n","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":28381,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":28381,"type":"exon","fmax":28699,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":28699,"type":"intron","fmax":29335,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29335,"type":"exon","fmax":29515,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29515,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":31339,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31339,"type":"exon","fmax":31582,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31582,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":38990,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38990,"type":"exon","fmax":39026,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":39026,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43697,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43697,"seqId":"V"}],"name":"B0348.4n.1","source":"WormBase","id":"Transcript:B0348.4n.1","fmin":20824,"type":"mRNA","fmax":43697,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4j","source":"WormBase","id":"CDS:B0348.4j","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4j","source":"WormBase","id":"CDS:B0348.4j","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4j","source":"WormBase","id":"CDS:B0348.4j","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4j","source":"WormBase","id":"CDS:B0348.4j","fmin":23561,"type":"CDS","fmax":23645,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4j","source":"WormBase","id":"CDS:B0348.4j","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4j","source":"WormBase","id":"CDS:B0348.4j","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4j","source":"WormBase","id":"CDS:B0348.4j","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4j","source":"WormBase","id":"CDS:B0348.4j","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4j","source":"WormBase","id":"CDS:B0348.4j","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4j","source":"WormBase","id":"CDS:B0348.4j","fmin":31339,"type":"CDS","fmax":31582,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4j","source":"WormBase","id":"CDS:B0348.4j","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4j","source":"WormBase","id":"CDS:B0348.4j","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4j","source":"WormBase","id":"CDS:B0348.4j","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4j","source":"WormBase","id":"CDS:B0348.4j","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4j","source":"WormBase","id":"CDS:B0348.4j","fmin":38990,"type":"CDS","fmax":39026,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4j","source":"WormBase","id":"CDS:B0348.4j","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4j","source":"WormBase","id":"CDS:B0348.4j","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4j","source":"WormBase","id":"CDS:B0348.4j","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4j","source":"WormBase","id":"CDS:B0348.4j","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4j","source":"WormBase","id":"CDS:B0348.4j","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23561,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23561,"type":"exon","fmax":23645,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23645,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":31339,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31339,"type":"exon","fmax":31582,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31582,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":38990,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38990,"type":"exon","fmax":39026,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":39026,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43697,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43697,"seqId":"V"}],"name":"B0348.4j.1","source":"WormBase","id":"Transcript:B0348.4j.1","fmin":20824,"type":"mRNA","fmax":43697,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4k","source":"WormBase","id":"CDS:B0348.4k","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4k","source":"WormBase","id":"CDS:B0348.4k","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4k","source":"WormBase","id":"CDS:B0348.4k","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4k","source":"WormBase","id":"CDS:B0348.4k","fmin":23561,"type":"CDS","fmax":23645,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4k","source":"WormBase","id":"CDS:B0348.4k","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4k","source":"WormBase","id":"CDS:B0348.4k","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4k","source":"WormBase","id":"CDS:B0348.4k","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4k","source":"WormBase","id":"CDS:B0348.4k","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4k","source":"WormBase","id":"CDS:B0348.4k","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4k","source":"WormBase","id":"CDS:B0348.4k","fmin":31339,"type":"CDS","fmax":31582,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4k","source":"WormBase","id":"CDS:B0348.4k","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4k","source":"WormBase","id":"CDS:B0348.4k","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4k","source":"WormBase","id":"CDS:B0348.4k","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4k","source":"WormBase","id":"CDS:B0348.4k","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4k","source":"WormBase","id":"CDS:B0348.4k","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4k","source":"WormBase","id":"CDS:B0348.4k","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4k","source":"WormBase","id":"CDS:B0348.4k","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4k","source":"WormBase","id":"CDS:B0348.4k","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4k","source":"WormBase","id":"CDS:B0348.4k","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23561,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23561,"type":"exon","fmax":23645,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23645,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":31339,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31339,"type":"exon","fmax":31582,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31582,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43687,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43687,"seqId":"V"}],"name":"B0348.4k.1","source":"WormBase","id":"Transcript:B0348.4k.1","fmin":20824,"type":"mRNA","fmax":43687,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4l","source":"WormBase","id":"CDS:B0348.4l","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4l","source":"WormBase","id":"CDS:B0348.4l","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4l","source":"WormBase","id":"CDS:B0348.4l","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4l","source":"WormBase","id":"CDS:B0348.4l","fmin":23561,"type":"CDS","fmax":23645,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4l","source":"WormBase","id":"CDS:B0348.4l","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4l","source":"WormBase","id":"CDS:B0348.4l","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4l","source":"WormBase","id":"CDS:B0348.4l","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4l","source":"WormBase","id":"CDS:B0348.4l","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4l","source":"WormBase","id":"CDS:B0348.4l","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4l","source":"WormBase","id":"CDS:B0348.4l","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4l","source":"WormBase","id":"CDS:B0348.4l","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4l","source":"WormBase","id":"CDS:B0348.4l","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4l","source":"WormBase","id":"CDS:B0348.4l","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4l","source":"WormBase","id":"CDS:B0348.4l","fmin":38990,"type":"CDS","fmax":39026,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4l","source":"WormBase","id":"CDS:B0348.4l","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4l","source":"WormBase","id":"CDS:B0348.4l","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4l","source":"WormBase","id":"CDS:B0348.4l","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4l","source":"WormBase","id":"CDS:B0348.4l","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4l","source":"WormBase","id":"CDS:B0348.4l","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23561,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23561,"type":"exon","fmax":23645,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23645,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":38990,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38990,"type":"exon","fmax":39026,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":39026,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43697,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43697,"seqId":"V"}],"name":"B0348.4l.1","source":"WormBase","id":"Transcript:B0348.4l.1","fmin":20824,"type":"mRNA","fmax":43697,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4x","source":"WormBase","id":"CDS:B0348.4x","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4x","source":"WormBase","id":"CDS:B0348.4x","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4x","source":"WormBase","id":"CDS:B0348.4x","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4x","source":"WormBase","id":"CDS:B0348.4x","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4x","source":"WormBase","id":"CDS:B0348.4x","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4x","source":"WormBase","id":"CDS:B0348.4x","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4x","source":"WormBase","id":"CDS:B0348.4x","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4x","source":"WormBase","id":"CDS:B0348.4x","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4x","source":"WormBase","id":"CDS:B0348.4x","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4x","source":"WormBase","id":"CDS:B0348.4x","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4x","source":"WormBase","id":"CDS:B0348.4x","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4x","source":"WormBase","id":"CDS:B0348.4x","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4x","source":"WormBase","id":"CDS:B0348.4x","fmin":38990,"type":"CDS","fmax":39026,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4x","source":"WormBase","id":"CDS:B0348.4x","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4x","source":"WormBase","id":"CDS:B0348.4x","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4x","source":"WormBase","id":"CDS:B0348.4x","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4x","source":"WormBase","id":"CDS:B0348.4x","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4x","source":"WormBase","id":"CDS:B0348.4x","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":38990,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38990,"type":"exon","fmax":39026,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":39026,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43697,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43697,"seqId":"V"}],"name":"B0348.4x.1","source":"WormBase","id":"Transcript:B0348.4x.1","fmin":20824,"type":"mRNA","fmax":43697,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4f","source":"WormBase","id":"CDS:B0348.4f","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4f","source":"WormBase","id":"CDS:B0348.4f","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4f","source":"WormBase","id":"CDS:B0348.4f","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4f","source":"WormBase","id":"CDS:B0348.4f","fmin":23561,"type":"CDS","fmax":23645,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4f","source":"WormBase","id":"CDS:B0348.4f","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4f","source":"WormBase","id":"CDS:B0348.4f","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4f","source":"WormBase","id":"CDS:B0348.4f","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4f","source":"WormBase","id":"CDS:B0348.4f","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4f","source":"WormBase","id":"CDS:B0348.4f","fmin":29335,"type":"CDS","fmax":29515,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4f","source":"WormBase","id":"CDS:B0348.4f","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4f","source":"WormBase","id":"CDS:B0348.4f","fmin":31339,"type":"CDS","fmax":31582,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4f","source":"WormBase","id":"CDS:B0348.4f","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4f","source":"WormBase","id":"CDS:B0348.4f","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4f","source":"WormBase","id":"CDS:B0348.4f","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4f","source":"WormBase","id":"CDS:B0348.4f","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4f","source":"WormBase","id":"CDS:B0348.4f","fmin":38990,"type":"CDS","fmax":39026,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4f","source":"WormBase","id":"CDS:B0348.4f","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4f","source":"WormBase","id":"CDS:B0348.4f","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4f","source":"WormBase","id":"CDS:B0348.4f","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4f","source":"WormBase","id":"CDS:B0348.4f","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4f","source":"WormBase","id":"CDS:B0348.4f","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23561,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23561,"type":"exon","fmax":23645,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23645,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":29335,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29335,"type":"exon","fmax":29515,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29515,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":31339,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31339,"type":"exon","fmax":31582,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31582,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":38990,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38990,"type":"exon","fmax":39026,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":39026,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43697,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43697,"seqId":"V"}],"name":"B0348.4f.1","source":"WormBase","id":"Transcript:B0348.4f.1","fmin":20824,"type":"mRNA","fmax":43697,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4q","source":"WormBase","id":"CDS:B0348.4q","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4q","source":"WormBase","id":"CDS:B0348.4q","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4q","source":"WormBase","id":"CDS:B0348.4q","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4q","source":"WormBase","id":"CDS:B0348.4q","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4q","source":"WormBase","id":"CDS:B0348.4q","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4q","source":"WormBase","id":"CDS:B0348.4q","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4q","source":"WormBase","id":"CDS:B0348.4q","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4q","source":"WormBase","id":"CDS:B0348.4q","fmin":28381,"type":"CDS","fmax":28699,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4q","source":"WormBase","id":"CDS:B0348.4q","fmin":29335,"type":"CDS","fmax":29515,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4q","source":"WormBase","id":"CDS:B0348.4q","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4q","source":"WormBase","id":"CDS:B0348.4q","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4q","source":"WormBase","id":"CDS:B0348.4q","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4q","source":"WormBase","id":"CDS:B0348.4q","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4q","source":"WormBase","id":"CDS:B0348.4q","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4q","source":"WormBase","id":"CDS:B0348.4q","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4q","source":"WormBase","id":"CDS:B0348.4q","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4q","source":"WormBase","id":"CDS:B0348.4q","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4q","source":"WormBase","id":"CDS:B0348.4q","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4q","source":"WormBase","id":"CDS:B0348.4q","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":28381,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":28381,"type":"exon","fmax":28699,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":28699,"type":"intron","fmax":29335,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29335,"type":"exon","fmax":29515,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29515,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43687,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43687,"seqId":"V"}],"name":"B0348.4q.1","source":"WormBase","id":"Transcript:B0348.4q.1","fmin":20824,"type":"mRNA","fmax":43687,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4y","source":"WormBase","id":"CDS:B0348.4y","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4y","source":"WormBase","id":"CDS:B0348.4y","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4y","source":"WormBase","id":"CDS:B0348.4y","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4y","source":"WormBase","id":"CDS:B0348.4y","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4y","source":"WormBase","id":"CDS:B0348.4y","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4y","source":"WormBase","id":"CDS:B0348.4y","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4y","source":"WormBase","id":"CDS:B0348.4y","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4y","source":"WormBase","id":"CDS:B0348.4y","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4y","source":"WormBase","id":"CDS:B0348.4y","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4y","source":"WormBase","id":"CDS:B0348.4y","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4y","source":"WormBase","id":"CDS:B0348.4y","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4y","source":"WormBase","id":"CDS:B0348.4y","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4y","source":"WormBase","id":"CDS:B0348.4y","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4y","source":"WormBase","id":"CDS:B0348.4y","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4y","source":"WormBase","id":"CDS:B0348.4y","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4y","source":"WormBase","id":"CDS:B0348.4y","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4y","source":"WormBase","id":"CDS:B0348.4y","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43687,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43687,"seqId":"V"}],"name":"B0348.4y.1","source":"WormBase","id":"Transcript:B0348.4y.1","fmin":20824,"type":"mRNA","fmax":43687,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4g","source":"WormBase","id":"CDS:B0348.4g","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4g","source":"WormBase","id":"CDS:B0348.4g","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4g","source":"WormBase","id":"CDS:B0348.4g","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4g","source":"WormBase","id":"CDS:B0348.4g","fmin":23561,"type":"CDS","fmax":23645,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4g","source":"WormBase","id":"CDS:B0348.4g","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4g","source":"WormBase","id":"CDS:B0348.4g","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4g","source":"WormBase","id":"CDS:B0348.4g","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4g","source":"WormBase","id":"CDS:B0348.4g","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4g","source":"WormBase","id":"CDS:B0348.4g","fmin":29335,"type":"CDS","fmax":29515,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4g","source":"WormBase","id":"CDS:B0348.4g","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4g","source":"WormBase","id":"CDS:B0348.4g","fmin":31339,"type":"CDS","fmax":31582,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4g","source":"WormBase","id":"CDS:B0348.4g","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4g","source":"WormBase","id":"CDS:B0348.4g","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4g","source":"WormBase","id":"CDS:B0348.4g","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4g","source":"WormBase","id":"CDS:B0348.4g","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4g","source":"WormBase","id":"CDS:B0348.4g","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4g","source":"WormBase","id":"CDS:B0348.4g","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4g","source":"WormBase","id":"CDS:B0348.4g","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4g","source":"WormBase","id":"CDS:B0348.4g","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4g","source":"WormBase","id":"CDS:B0348.4g","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23561,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23561,"type":"exon","fmax":23645,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23645,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":29335,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29335,"type":"exon","fmax":29515,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29515,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":31339,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31339,"type":"exon","fmax":31582,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31582,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43687,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43687,"seqId":"V"}],"name":"B0348.4g.1","source":"WormBase","id":"Transcript:B0348.4g.1","fmin":20824,"type":"mRNA","fmax":43687,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4u","source":"WormBase","id":"CDS:B0348.4u","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4u","source":"WormBase","id":"CDS:B0348.4u","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4u","source":"WormBase","id":"CDS:B0348.4u","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4u","source":"WormBase","id":"CDS:B0348.4u","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4u","source":"WormBase","id":"CDS:B0348.4u","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4u","source":"WormBase","id":"CDS:B0348.4u","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4u","source":"WormBase","id":"CDS:B0348.4u","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4u","source":"WormBase","id":"CDS:B0348.4u","fmin":29335,"type":"CDS","fmax":29515,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4u","source":"WormBase","id":"CDS:B0348.4u","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4u","source":"WormBase","id":"CDS:B0348.4u","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4u","source":"WormBase","id":"CDS:B0348.4u","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4u","source":"WormBase","id":"CDS:B0348.4u","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4u","source":"WormBase","id":"CDS:B0348.4u","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4u","source":"WormBase","id":"CDS:B0348.4u","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4u","source":"WormBase","id":"CDS:B0348.4u","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4u","source":"WormBase","id":"CDS:B0348.4u","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4u","source":"WormBase","id":"CDS:B0348.4u","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4u","source":"WormBase","id":"CDS:B0348.4u","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":29335,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29335,"type":"exon","fmax":29515,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29515,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43687,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43687,"seqId":"V"}],"name":"B0348.4u.1","source":"WormBase","id":"Transcript:B0348.4u.1","fmin":20824,"type":"mRNA","fmax":43687,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4i","source":"WormBase","id":"CDS:B0348.4i","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4i","source":"WormBase","id":"CDS:B0348.4i","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4i","source":"WormBase","id":"CDS:B0348.4i","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4i","source":"WormBase","id":"CDS:B0348.4i","fmin":23561,"type":"CDS","fmax":23645,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4i","source":"WormBase","id":"CDS:B0348.4i","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4i","source":"WormBase","id":"CDS:B0348.4i","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4i","source":"WormBase","id":"CDS:B0348.4i","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4i","source":"WormBase","id":"CDS:B0348.4i","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4i","source":"WormBase","id":"CDS:B0348.4i","fmin":29335,"type":"CDS","fmax":29515,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4i","source":"WormBase","id":"CDS:B0348.4i","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4i","source":"WormBase","id":"CDS:B0348.4i","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4i","source":"WormBase","id":"CDS:B0348.4i","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4i","source":"WormBase","id":"CDS:B0348.4i","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4i","source":"WormBase","id":"CDS:B0348.4i","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4i","source":"WormBase","id":"CDS:B0348.4i","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4i","source":"WormBase","id":"CDS:B0348.4i","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4i","source":"WormBase","id":"CDS:B0348.4i","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4i","source":"WormBase","id":"CDS:B0348.4i","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4i","source":"WormBase","id":"CDS:B0348.4i","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23561,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23561,"type":"exon","fmax":23645,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23645,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":29335,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29335,"type":"exon","fmax":29515,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29515,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43687,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43687,"seqId":"V"}],"name":"B0348.4i.1","source":"WormBase","id":"Transcript:B0348.4i.1","fmin":20824,"type":"mRNA","fmax":43687,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4s","source":"WormBase","id":"CDS:B0348.4s","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4s","source":"WormBase","id":"CDS:B0348.4s","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4s","source":"WormBase","id":"CDS:B0348.4s","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4s","source":"WormBase","id":"CDS:B0348.4s","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4s","source":"WormBase","id":"CDS:B0348.4s","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4s","source":"WormBase","id":"CDS:B0348.4s","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4s","source":"WormBase","id":"CDS:B0348.4s","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4s","source":"WormBase","id":"CDS:B0348.4s","fmin":29335,"type":"CDS","fmax":29515,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4s","source":"WormBase","id":"CDS:B0348.4s","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4s","source":"WormBase","id":"CDS:B0348.4s","fmin":31339,"type":"CDS","fmax":31582,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4s","source":"WormBase","id":"CDS:B0348.4s","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4s","source":"WormBase","id":"CDS:B0348.4s","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4s","source":"WormBase","id":"CDS:B0348.4s","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4s","source":"WormBase","id":"CDS:B0348.4s","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4s","source":"WormBase","id":"CDS:B0348.4s","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4s","source":"WormBase","id":"CDS:B0348.4s","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4s","source":"WormBase","id":"CDS:B0348.4s","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4s","source":"WormBase","id":"CDS:B0348.4s","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4s","source":"WormBase","id":"CDS:B0348.4s","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":29335,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29335,"type":"exon","fmax":29515,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29515,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":31339,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31339,"type":"exon","fmax":31582,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31582,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43687,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43687,"seqId":"V"}],"name":"B0348.4s.1","source":"WormBase","id":"Transcript:B0348.4s.1","fmin":20824,"type":"mRNA","fmax":43687,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4a","source":"WormBase","id":"CDS:B0348.4a","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4a","source":"WormBase","id":"CDS:B0348.4a","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4a","source":"WormBase","id":"CDS:B0348.4a","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4a","source":"WormBase","id":"CDS:B0348.4a","fmin":23561,"type":"CDS","fmax":23645,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4a","source":"WormBase","id":"CDS:B0348.4a","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4a","source":"WormBase","id":"CDS:B0348.4a","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4a","source":"WormBase","id":"CDS:B0348.4a","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4a","source":"WormBase","id":"CDS:B0348.4a","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4a","source":"WormBase","id":"CDS:B0348.4a","fmin":28381,"type":"CDS","fmax":28699,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4a","source":"WormBase","id":"CDS:B0348.4a","fmin":29335,"type":"CDS","fmax":29515,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4a","source":"WormBase","id":"CDS:B0348.4a","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4a","source":"WormBase","id":"CDS:B0348.4a","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4a","source":"WormBase","id":"CDS:B0348.4a","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4a","source":"WormBase","id":"CDS:B0348.4a","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4a","source":"WormBase","id":"CDS:B0348.4a","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4a","source":"WormBase","id":"CDS:B0348.4a","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4a","source":"WormBase","id":"CDS:B0348.4a","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4a","source":"WormBase","id":"CDS:B0348.4a","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4a","source":"WormBase","id":"CDS:B0348.4a","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4a","source":"WormBase","id":"CDS:B0348.4a","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23561,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23561,"type":"exon","fmax":23645,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23645,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":28381,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":28381,"type":"exon","fmax":28699,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":28699,"type":"intron","fmax":29335,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29335,"type":"exon","fmax":29515,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29515,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43687,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43687,"seqId":"V"}],"name":"B0348.4a.1","source":"WormBase","id":"Transcript:B0348.4a.1","fmin":20824,"type":"mRNA","fmax":43687,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4b","source":"WormBase","id":"CDS:B0348.4b","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4b","source":"WormBase","id":"CDS:B0348.4b","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4b","source":"WormBase","id":"CDS:B0348.4b","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4b","source":"WormBase","id":"CDS:B0348.4b","fmin":23561,"type":"CDS","fmax":23645,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4b","source":"WormBase","id":"CDS:B0348.4b","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4b","source":"WormBase","id":"CDS:B0348.4b","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4b","source":"WormBase","id":"CDS:B0348.4b","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4b","source":"WormBase","id":"CDS:B0348.4b","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4b","source":"WormBase","id":"CDS:B0348.4b","fmin":28381,"type":"CDS","fmax":28699,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4b","source":"WormBase","id":"CDS:B0348.4b","fmin":29335,"type":"CDS","fmax":29515,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4b","source":"WormBase","id":"CDS:B0348.4b","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4b","source":"WormBase","id":"CDS:B0348.4b","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4b","source":"WormBase","id":"CDS:B0348.4b","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4b","source":"WormBase","id":"CDS:B0348.4b","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4b","source":"WormBase","id":"CDS:B0348.4b","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4b","source":"WormBase","id":"CDS:B0348.4b","fmin":38990,"type":"CDS","fmax":39026,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4b","source":"WormBase","id":"CDS:B0348.4b","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4b","source":"WormBase","id":"CDS:B0348.4b","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4b","source":"WormBase","id":"CDS:B0348.4b","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4b","source":"WormBase","id":"CDS:B0348.4b","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4b","source":"WormBase","id":"CDS:B0348.4b","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23561,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23561,"type":"exon","fmax":23645,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23645,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":28381,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":28381,"type":"exon","fmax":28699,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":28699,"type":"intron","fmax":29335,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29335,"type":"exon","fmax":29515,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29515,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":38990,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38990,"type":"exon","fmax":39026,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":39026,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43697,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43697,"seqId":"V"}],"name":"B0348.4b.1","source":"WormBase","id":"Transcript:B0348.4b.1","fmin":20824,"type":"mRNA","fmax":43697,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4r","source":"WormBase","id":"CDS:B0348.4r","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4r","source":"WormBase","id":"CDS:B0348.4r","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4r","source":"WormBase","id":"CDS:B0348.4r","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4r","source":"WormBase","id":"CDS:B0348.4r","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4r","source":"WormBase","id":"CDS:B0348.4r","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4r","source":"WormBase","id":"CDS:B0348.4r","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4r","source":"WormBase","id":"CDS:B0348.4r","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4r","source":"WormBase","id":"CDS:B0348.4r","fmin":29335,"type":"CDS","fmax":29515,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4r","source":"WormBase","id":"CDS:B0348.4r","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4r","source":"WormBase","id":"CDS:B0348.4r","fmin":31339,"type":"CDS","fmax":31582,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4r","source":"WormBase","id":"CDS:B0348.4r","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4r","source":"WormBase","id":"CDS:B0348.4r","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4r","source":"WormBase","id":"CDS:B0348.4r","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4r","source":"WormBase","id":"CDS:B0348.4r","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4r","source":"WormBase","id":"CDS:B0348.4r","fmin":38990,"type":"CDS","fmax":39026,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4r","source":"WormBase","id":"CDS:B0348.4r","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4r","source":"WormBase","id":"CDS:B0348.4r","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4r","source":"WormBase","id":"CDS:B0348.4r","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4r","source":"WormBase","id":"CDS:B0348.4r","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4r","source":"WormBase","id":"CDS:B0348.4r","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":29335,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29335,"type":"exon","fmax":29515,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29515,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":31339,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31339,"type":"exon","fmax":31582,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31582,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":38990,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38990,"type":"exon","fmax":39026,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":39026,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43697,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43697,"seqId":"V"}],"name":"B0348.4r.1","source":"WormBase","id":"Transcript:B0348.4r.1","fmin":20824,"type":"mRNA","fmax":43697,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4v","source":"WormBase","id":"CDS:B0348.4v","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4v","source":"WormBase","id":"CDS:B0348.4v","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4v","source":"WormBase","id":"CDS:B0348.4v","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4v","source":"WormBase","id":"CDS:B0348.4v","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4v","source":"WormBase","id":"CDS:B0348.4v","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4v","source":"WormBase","id":"CDS:B0348.4v","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4v","source":"WormBase","id":"CDS:B0348.4v","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4v","source":"WormBase","id":"CDS:B0348.4v","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4v","source":"WormBase","id":"CDS:B0348.4v","fmin":31339,"type":"CDS","fmax":31582,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4v","source":"WormBase","id":"CDS:B0348.4v","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4v","source":"WormBase","id":"CDS:B0348.4v","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4v","source":"WormBase","id":"CDS:B0348.4v","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4v","source":"WormBase","id":"CDS:B0348.4v","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4v","source":"WormBase","id":"CDS:B0348.4v","fmin":38990,"type":"CDS","fmax":39026,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4v","source":"WormBase","id":"CDS:B0348.4v","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4v","source":"WormBase","id":"CDS:B0348.4v","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4v","source":"WormBase","id":"CDS:B0348.4v","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4v","source":"WormBase","id":"CDS:B0348.4v","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4v","source":"WormBase","id":"CDS:B0348.4v","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":31339,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31339,"type":"exon","fmax":31582,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31582,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":38990,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38990,"type":"exon","fmax":39026,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":39026,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43697,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43697,"seqId":"V"}],"name":"B0348.4v.1","source":"WormBase","id":"Transcript:B0348.4v.1","fmin":20824,"type":"mRNA","fmax":43697,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4w","source":"WormBase","id":"CDS:B0348.4w","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4w","source":"WormBase","id":"CDS:B0348.4w","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4w","source":"WormBase","id":"CDS:B0348.4w","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4w","source":"WormBase","id":"CDS:B0348.4w","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4w","source":"WormBase","id":"CDS:B0348.4w","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4w","source":"WormBase","id":"CDS:B0348.4w","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4w","source":"WormBase","id":"CDS:B0348.4w","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4w","source":"WormBase","id":"CDS:B0348.4w","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4w","source":"WormBase","id":"CDS:B0348.4w","fmin":31339,"type":"CDS","fmax":31582,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4w","source":"WormBase","id":"CDS:B0348.4w","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4w","source":"WormBase","id":"CDS:B0348.4w","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4w","source":"WormBase","id":"CDS:B0348.4w","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4w","source":"WormBase","id":"CDS:B0348.4w","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4w","source":"WormBase","id":"CDS:B0348.4w","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4w","source":"WormBase","id":"CDS:B0348.4w","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4w","source":"WormBase","id":"CDS:B0348.4w","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4w","source":"WormBase","id":"CDS:B0348.4w","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4w","source":"WormBase","id":"CDS:B0348.4w","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":31339,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31339,"type":"exon","fmax":31582,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31582,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43687,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43687,"seqId":"V"}],"name":"B0348.4w.1","source":"WormBase","id":"Transcript:B0348.4w.1","fmin":20824,"type":"mRNA","fmax":43687,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4t","source":"WormBase","id":"CDS:B0348.4t","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4t","source":"WormBase","id":"CDS:B0348.4t","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4t","source":"WormBase","id":"CDS:B0348.4t","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4t","source":"WormBase","id":"CDS:B0348.4t","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4t","source":"WormBase","id":"CDS:B0348.4t","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4t","source":"WormBase","id":"CDS:B0348.4t","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4t","source":"WormBase","id":"CDS:B0348.4t","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4t","source":"WormBase","id":"CDS:B0348.4t","fmin":29335,"type":"CDS","fmax":29515,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4t","source":"WormBase","id":"CDS:B0348.4t","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4t","source":"WormBase","id":"CDS:B0348.4t","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4t","source":"WormBase","id":"CDS:B0348.4t","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4t","source":"WormBase","id":"CDS:B0348.4t","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4t","source":"WormBase","id":"CDS:B0348.4t","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4t","source":"WormBase","id":"CDS:B0348.4t","fmin":38990,"type":"CDS","fmax":39026,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4t","source":"WormBase","id":"CDS:B0348.4t","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4t","source":"WormBase","id":"CDS:B0348.4t","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4t","source":"WormBase","id":"CDS:B0348.4t","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4t","source":"WormBase","id":"CDS:B0348.4t","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4t","source":"WormBase","id":"CDS:B0348.4t","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":29335,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29335,"type":"exon","fmax":29515,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29515,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":38990,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38990,"type":"exon","fmax":39026,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":39026,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43697,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43697,"seqId":"V"}],"name":"B0348.4t.1","source":"WormBase","id":"Transcript:B0348.4t.1","fmin":20824,"type":"mRNA","fmax":43697,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4h","source":"WormBase","id":"CDS:B0348.4h","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4h","source":"WormBase","id":"CDS:B0348.4h","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4h","source":"WormBase","id":"CDS:B0348.4h","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4h","source":"WormBase","id":"CDS:B0348.4h","fmin":23561,"type":"CDS","fmax":23645,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4h","source":"WormBase","id":"CDS:B0348.4h","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4h","source":"WormBase","id":"CDS:B0348.4h","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4h","source":"WormBase","id":"CDS:B0348.4h","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4h","source":"WormBase","id":"CDS:B0348.4h","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4h","source":"WormBase","id":"CDS:B0348.4h","fmin":29335,"type":"CDS","fmax":29515,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4h","source":"WormBase","id":"CDS:B0348.4h","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4h","source":"WormBase","id":"CDS:B0348.4h","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4h","source":"WormBase","id":"CDS:B0348.4h","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4h","source":"WormBase","id":"CDS:B0348.4h","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4h","source":"WormBase","id":"CDS:B0348.4h","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4h","source":"WormBase","id":"CDS:B0348.4h","fmin":38990,"type":"CDS","fmax":39026,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4h","source":"WormBase","id":"CDS:B0348.4h","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4h","source":"WormBase","id":"CDS:B0348.4h","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4h","source":"WormBase","id":"CDS:B0348.4h","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4h","source":"WormBase","id":"CDS:B0348.4h","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4h","source":"WormBase","id":"CDS:B0348.4h","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23561,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23561,"type":"exon","fmax":23645,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23645,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":29335,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29335,"type":"exon","fmax":29515,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29515,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":38990,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38990,"type":"exon","fmax":39026,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":39026,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43697,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43697,"seqId":"V"}],"name":"B0348.4h.1","source":"WormBase","id":"Transcript:B0348.4h.1","fmin":20824,"type":"mRNA","fmax":43697,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4o","source":"WormBase","id":"CDS:B0348.4o","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4o","source":"WormBase","id":"CDS:B0348.4o","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4o","source":"WormBase","id":"CDS:B0348.4o","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4o","source":"WormBase","id":"CDS:B0348.4o","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4o","source":"WormBase","id":"CDS:B0348.4o","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4o","source":"WormBase","id":"CDS:B0348.4o","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4o","source":"WormBase","id":"CDS:B0348.4o","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4o","source":"WormBase","id":"CDS:B0348.4o","fmin":28381,"type":"CDS","fmax":28699,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4o","source":"WormBase","id":"CDS:B0348.4o","fmin":29335,"type":"CDS","fmax":29515,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4o","source":"WormBase","id":"CDS:B0348.4o","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4o","source":"WormBase","id":"CDS:B0348.4o","fmin":31339,"type":"CDS","fmax":31582,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4o","source":"WormBase","id":"CDS:B0348.4o","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4o","source":"WormBase","id":"CDS:B0348.4o","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4o","source":"WormBase","id":"CDS:B0348.4o","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4o","source":"WormBase","id":"CDS:B0348.4o","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4o","source":"WormBase","id":"CDS:B0348.4o","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4o","source":"WormBase","id":"CDS:B0348.4o","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4o","source":"WormBase","id":"CDS:B0348.4o","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4o","source":"WormBase","id":"CDS:B0348.4o","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4o","source":"WormBase","id":"CDS:B0348.4o","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":28381,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":28381,"type":"exon","fmax":28699,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":28699,"type":"intron","fmax":29335,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29335,"type":"exon","fmax":29515,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29515,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":31339,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31339,"type":"exon","fmax":31582,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31582,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43687,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43687,"seqId":"V"}],"name":"B0348.4o.1","source":"WormBase","id":"Transcript:B0348.4o.1","fmin":20824,"type":"mRNA","fmax":43687,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4e","source":"WormBase","id":"CDS:B0348.4e","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4e","source":"WormBase","id":"CDS:B0348.4e","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4e","source":"WormBase","id":"CDS:B0348.4e","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4e","source":"WormBase","id":"CDS:B0348.4e","fmin":23561,"type":"CDS","fmax":23645,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4e","source":"WormBase","id":"CDS:B0348.4e","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4e","source":"WormBase","id":"CDS:B0348.4e","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4e","source":"WormBase","id":"CDS:B0348.4e","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4e","source":"WormBase","id":"CDS:B0348.4e","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4e","source":"WormBase","id":"CDS:B0348.4e","fmin":28381,"type":"CDS","fmax":28699,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4e","source":"WormBase","id":"CDS:B0348.4e","fmin":29335,"type":"CDS","fmax":29515,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4e","source":"WormBase","id":"CDS:B0348.4e","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4e","source":"WormBase","id":"CDS:B0348.4e","fmin":31339,"type":"CDS","fmax":31582,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4e","source":"WormBase","id":"CDS:B0348.4e","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4e","source":"WormBase","id":"CDS:B0348.4e","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4e","source":"WormBase","id":"CDS:B0348.4e","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4e","source":"WormBase","id":"CDS:B0348.4e","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4e","source":"WormBase","id":"CDS:B0348.4e","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4e","source":"WormBase","id":"CDS:B0348.4e","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4e","source":"WormBase","id":"CDS:B0348.4e","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4e","source":"WormBase","id":"CDS:B0348.4e","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4e","source":"WormBase","id":"CDS:B0348.4e","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23561,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23561,"type":"exon","fmax":23645,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23645,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":28381,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":28381,"type":"exon","fmax":28699,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":28699,"type":"intron","fmax":29335,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29335,"type":"exon","fmax":29515,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29515,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":31339,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31339,"type":"exon","fmax":31582,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31582,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43687,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43687,"seqId":"V"}],"name":"B0348.4e.1","source":"WormBase","id":"Transcript:B0348.4e.1","fmin":20824,"type":"mRNA","fmax":43687,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4c","source":"WormBase","id":"CDS:B0348.4c","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4c","source":"WormBase","id":"CDS:B0348.4c","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4c","source":"WormBase","id":"CDS:B0348.4c","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4c","source":"WormBase","id":"CDS:B0348.4c","fmin":23561,"type":"CDS","fmax":23645,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4c","source":"WormBase","id":"CDS:B0348.4c","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4c","source":"WormBase","id":"CDS:B0348.4c","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4c","source":"WormBase","id":"CDS:B0348.4c","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4c","source":"WormBase","id":"CDS:B0348.4c","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4c","source":"WormBase","id":"CDS:B0348.4c","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4c","source":"WormBase","id":"CDS:B0348.4c","fmin":30354,"type":"CDS","fmax":30441,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23561,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23561,"type":"exon","fmax":23645,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23645,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":30354,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30354,"type":"exon","fmax":30703,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30441,"type":"three_prime_UTR","fmax":30703,"seqId":"V"}],"name":"B0348.4c.1","source":"WormBase","id":"Transcript:B0348.4c.1","fmin":20824,"type":"mRNA","fmax":30703,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":23561,"type":"CDS","fmax":23645,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":28381,"type":"CDS","fmax":28699,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":29335,"type":"CDS","fmax":29515,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":31339,"type":"CDS","fmax":31582,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":38990,"type":"CDS","fmax":39026,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4d","source":"WormBase","id":"CDS:B0348.4d","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23561,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23561,"type":"exon","fmax":23645,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23645,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":28381,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":28381,"type":"exon","fmax":28699,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":28699,"type":"intron","fmax":29335,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29335,"type":"exon","fmax":29515,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29515,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":31339,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31339,"type":"exon","fmax":31582,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":31582,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":38990,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38990,"type":"exon","fmax":39026,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":39026,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43697,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43697,"seqId":"V"}],"name":"B0348.4d.1","source":"WormBase","id":"Transcript:B0348.4d.1","fmin":20824,"type":"mRNA","fmax":43697,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4p","source":"WormBase","id":"CDS:B0348.4p","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4p","source":"WormBase","id":"CDS:B0348.4p","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4p","source":"WormBase","id":"CDS:B0348.4p","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4p","source":"WormBase","id":"CDS:B0348.4p","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4p","source":"WormBase","id":"CDS:B0348.4p","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4p","source":"WormBase","id":"CDS:B0348.4p","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4p","source":"WormBase","id":"CDS:B0348.4p","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4p","source":"WormBase","id":"CDS:B0348.4p","fmin":28381,"type":"CDS","fmax":28699,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4p","source":"WormBase","id":"CDS:B0348.4p","fmin":29335,"type":"CDS","fmax":29515,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4p","source":"WormBase","id":"CDS:B0348.4p","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4p","source":"WormBase","id":"CDS:B0348.4p","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4p","source":"WormBase","id":"CDS:B0348.4p","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4p","source":"WormBase","id":"CDS:B0348.4p","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4p","source":"WormBase","id":"CDS:B0348.4p","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4p","source":"WormBase","id":"CDS:B0348.4p","fmin":38990,"type":"CDS","fmax":39026,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4p","source":"WormBase","id":"CDS:B0348.4p","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4p","source":"WormBase","id":"CDS:B0348.4p","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4p","source":"WormBase","id":"CDS:B0348.4p","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4p","source":"WormBase","id":"CDS:B0348.4p","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4p","source":"WormBase","id":"CDS:B0348.4p","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":28381,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":28381,"type":"exon","fmax":28699,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":28699,"type":"intron","fmax":29335,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29335,"type":"exon","fmax":29515,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":29515,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":38990,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38990,"type":"exon","fmax":39026,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":39026,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43697,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43697,"seqId":"V"}],"name":"B0348.4p.1","source":"WormBase","id":"Transcript:B0348.4p.1","fmin":20824,"type":"mRNA","fmax":43697,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":20824,"type":"exon","fmax":21090,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":20824,"type":"five_prime_UTR","fmax":21006,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4m","source":"WormBase","id":"CDS:B0348.4m","fmin":21006,"type":"CDS","fmax":21090,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4m","source":"WormBase","id":"CDS:B0348.4m","fmin":22494,"type":"CDS","fmax":22856,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4m","source":"WormBase","id":"CDS:B0348.4m","fmin":23158,"type":"CDS","fmax":23294,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4m","source":"WormBase","id":"CDS:B0348.4m","fmin":23561,"type":"CDS","fmax":23645,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4m","source":"WormBase","id":"CDS:B0348.4m","fmin":23792,"type":"CDS","fmax":23896,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4m","source":"WormBase","id":"CDS:B0348.4m","fmin":24787,"type":"CDS","fmax":25259,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4m","source":"WormBase","id":"CDS:B0348.4m","fmin":25964,"type":"CDS","fmax":26129,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4m","source":"WormBase","id":"CDS:B0348.4m","fmin":26193,"type":"CDS","fmax":26380,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4m","source":"WormBase","id":"CDS:B0348.4m","fmin":30199,"type":"CDS","fmax":30297,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4m","source":"WormBase","id":"CDS:B0348.4m","fmin":35014,"type":"CDS","fmax":35156,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4m","source":"WormBase","id":"CDS:B0348.4m","fmin":36117,"type":"CDS","fmax":36855,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.4m","source":"WormBase","id":"CDS:B0348.4m","fmin":36917,"type":"CDS","fmax":37213,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4m","source":"WormBase","id":"CDS:B0348.4m","fmin":38137,"type":"CDS","fmax":38269,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4m","source":"WormBase","id":"CDS:B0348.4m","fmin":40453,"type":"CDS","fmax":40723,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4m","source":"WormBase","id":"CDS:B0348.4m","fmin":41728,"type":"CDS","fmax":41844,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4m","source":"WormBase","id":"CDS:B0348.4m","fmin":42039,"type":"CDS","fmax":42166,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.4m","source":"WormBase","id":"CDS:B0348.4m","fmin":42213,"type":"CDS","fmax":42317,"seqId":"V"},{"phase":1,"strand":1,"name":"B0348.4m","source":"WormBase","id":"CDS:B0348.4m","fmin":43025,"type":"CDS","fmax":43170,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":21090,"type":"intron","fmax":22494,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22494,"type":"exon","fmax":22856,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":22856,"type":"intron","fmax":23158,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23158,"type":"exon","fmax":23294,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23294,"type":"intron","fmax":23561,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23561,"type":"exon","fmax":23645,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23645,"type":"intron","fmax":23792,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23792,"type":"exon","fmax":23896,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":23896,"type":"intron","fmax":24787,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":24787,"type":"exon","fmax":25259,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25259,"type":"intron","fmax":25964,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":25964,"type":"exon","fmax":26129,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26129,"type":"intron","fmax":26193,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26193,"type":"exon","fmax":26380,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":26380,"type":"intron","fmax":30199,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30199,"type":"exon","fmax":30297,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":30297,"type":"intron","fmax":35014,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35014,"type":"exon","fmax":35156,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":35156,"type":"intron","fmax":36117,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36117,"type":"exon","fmax":36855,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36855,"type":"intron","fmax":36917,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":36917,"type":"exon","fmax":37213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":37213,"type":"intron","fmax":38137,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38137,"type":"exon","fmax":38269,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":38269,"type":"intron","fmax":40453,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40453,"type":"exon","fmax":40723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":40723,"type":"intron","fmax":41728,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41728,"type":"exon","fmax":41844,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":41844,"type":"intron","fmax":42039,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42039,"type":"exon","fmax":42166,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42166,"type":"intron","fmax":42213,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42213,"type":"exon","fmax":42317,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":42317,"type":"intron","fmax":43025,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43025,"type":"exon","fmax":43687,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":43170,"type":"three_prime_UTR","fmax":43687,"seqId":"V"}],"name":"B0348.4m.1","source":"WormBase","id":"Transcript:B0348.4m.1","fmin":20824,"type":"mRNA","fmax":43687,"seqId":"V"},[]],"name":"egl-8","source":"WormBase","id":"http://localhost:8080/apollo/track/worm/All Genes/V/egl-8.json","fmin":20824,"type":"gene","fmax":43697,"seqId":"V"},{"strand":1,"children":[{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":45309,"type":"five_prime_UTR","fmax":45466,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":45309,"type":"exon","fmax":45499,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.2a","source":"WormBase","id":"CDS:B0348.2a","fmin":45466,"type":"CDS","fmax":45499,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.2a","source":"WormBase","id":"CDS:B0348.2a","fmin":45598,"type":"CDS","fmax":45646,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.2a","source":"WormBase","id":"CDS:B0348.2a","fmin":46700,"type":"CDS","fmax":46880,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":45499,"type":"intron","fmax":45598,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":45598,"type":"exon","fmax":45646,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":45646,"type":"intron","fmax":46700,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":46700,"type":"exon","fmax":46916,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":46880,"type":"three_prime_UTR","fmax":46916,"seqId":"V"}],"name":"B0348.2a.1","source":"WormBase","id":"Transcript:B0348.2a.1","fmin":45309,"type":"mRNA","fmax":46916,"seqId":"V"},{"strand":1,"children":[{"strand":1,"source":"WormBase","fmin":45309,"type":"five_prime_UTR","fmax":45606,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":45309,"type":"exon","fmax":45646,"seqId":"V"},{"phase":0,"strand":1,"name":"B0348.2b","source":"WormBase","id":"CDS:B0348.2b","fmin":45606,"type":"CDS","fmax":45646,"seqId":"V"},{"phase":2,"strand":1,"name":"B0348.2b","source":"WormBase","id":"CDS:B0348.2b","fmin":46700,"type":"CDS","fmax":46723,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":45646,"type":"intron","fmax":46700,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":46700,"type":"exon","fmax":46916,"seqId":"V"},{"strand":1,"source":"WormBase","fmin":46723,"type":"three_prime_UTR","fmax":46916,"seqId":"V"}],"name":"B0348.2b.1","source":"WormBase","id":"Transcript:B0348.2b.1","fmin":45309,"type":"mRNA","fmax":46916,"seqId":"V"}],"name":"B0348.2","source":"WormBase","id":"http://localhost:8080/apollo/track/worm/All Genes/V/B0348.2.json","fmin":45309,"type":"gene","fmax":46916,"seqId":"V"}];
     }
 
-    GetFakeWormVariantData(){
+    GetFakeWormVariantDataEgl8(){
       // http://localhost:8080/apollo/vcf/worm/Phenotypic%20Variants/V:17405..47604.json?includeGenotypes=false&ignoreCache=true
       return [{"symbol":{"meta":{"number":[1],"description":["The human readable name of the allele"],"id":["symbol"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"tm774\""]},"alternative_alleles":{"meta":{"description":"VCF ALT field, list of alternate non-reference alleles called on at least one of the samples"},"values":"C"},"geneLevelConsequence":{"meta":{"number":[-1],"description":["VEP consequence of the variant"],"id":["geneLevelConsequence"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":["\"intron_variant\""]},"hgvs_nomenclature":{"meta":{"number":[1],"description":["the HGVS name of the allele"],"id":["hgvs_nomenclature"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"NC_003283.11:g.23999_24591del\""]},"impact":{"meta":{"number":[1],"description":["Variant impact scale"],"id":["impact"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"MODIFIER\""]},"description":"deletion CTAACAGAAAATGTACTATTACACCGATTATTAGTACTTATAATTTTTTTTTTGAAAGTTGGAAAATATTTTAAAGTTTGTAAAAATTCTATAAATAATATTGAATATTATAACGGTAAAATATAATATTTATAGAATTTTTTCAATTTTCAAAAAAGTACTGTTTGTCATTTGTTTTTGTTTGTTAGAATACCTTTTAATAGCGGTACCGAAATCTGGGAAATATTTTCAAATGACTCCAAATTTTCGCCTGATTCCGAATATCTATGTGAAAAAAAGTGAAAAAAAACATCCCTTATTTTATATTTCATTTCCAATTCCCAATTTCATTTGTGCCCGCATTACTTTTTTCAAACGCGCGCCCAAATAAATTTTCATTAGAGCGCGTTTGCTTTGTGTCGATTTACGGGAGCTCTCCATTTACAAAAAAAATTAACGCTTTTTTTAAACAGTTTTTAAGCGAGTTTCCTACTTTTTTAATCGATTTCAGCGGAACGGTTTTTCTTGAAATATCGTTTTTAAGGTCAATTGATATGATAATAAAATATATTGTTGACCTTAAAACACATTAAAAAAACTTCGACAAGAACCGAC > C","type":"deletion","symbol_text":{"meta":{"number":[1],"description":["Another human readable representation of the allele"],"id":["symbol_text"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"tm774\""]},"fmax":24591,"reference_allele":"CTAACAGAAAATGTACTATTACACCGATTATTAGTACTTATAATTTTTTTTTTGAAAGTTGGAAAATATTTTAAAGTTTGTAAAAATTCTATAAATAATATTGAATATTATAACGGTAAAATATAATATTTATAGAATTTTTTCAATTTTCAAAAAAGTACTGTTTGTCATTTGTTTTTGTTTGTTAGAATACCTTTTAATAGCGGTACCGAAATCTGGGAAATATTTTCAAATGACTCCAAATTTTCGCCTGATTCCGAATATCTATGTGAAAAAAAGTGAAAAAAAACATCCCTTATTTTATATTTCATTTCCAATTCCCAATTTCATTTGTGCCCGCATTACTTTTTTCAAACGCGCGCCCAAATAAATTTTCATTAGAGCGCGTTTGCTTTGTGTCGATTTACGGGAGCTCTCCATTTACAAAAAAAATTAACGCTTTTTTTAAACAGTTTTTAAGCGAGTTTCCTACTTTTTTAATCGATTTCAGCGGAACGGTTTTTCTTGAAATATCGTTTTTAAGGTCAATTGATATGATAATAAAATATATTGTTGACCTTAAAACACATTAAAAAAACTTCGACAAGAACCGAC","name":"NC_003283.11:g.23999_24591del","allele_of_genes":{"meta":{"number":[1],"description":["The genes that the Allele is located on"],"id":["allele_of_genes"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"WB:WBGene00001177\""]},"fmin":23997,"seqId":"V","uniqueID":"NC_003283.11:g.23999_24591del","alleles":{"meta":{"number":[-1],"description":["The alleles of the variant"],"id":["alleles"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":["\"WB:WBVar00249802\""]}},{"symbol":{"meta":{"number":[1],"description":["The human readable name of the allele"],"id":["symbol"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"nj77\""]},"alternative_alleles":{"meta":{"description":"VCF ALT field, list of alternate non-reference alleles called on at least one of the samples"},"values":"A"},"geneLevelConsequence":{"meta":{"number":[-1],"description":["VEP consequence of the variant"],"id":["geneLevelConsequence"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":["\"splice_donor_variant\""]},"hgvs_nomenclature":{"meta":{"number":[1],"description":["the HGVS name of the allele"],"id":["hgvs_nomenclature"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"NC_003283.11:g.26381G>A\""]},"impact":{"meta":{"number":[1],"description":["Variant impact scale"],"id":["impact"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"HIGH\""]},"description":"SNV G > A","type":"SNV","symbol_text":{"meta":{"number":[1],"description":["Another human readable representation of the allele"],"id":["symbol_text"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"nj77\""]},"fmax":26381,"reference_allele":"G","name":"NC_003283.11:g.26381G>A","allele_of_genes":{"meta":{"number":[1],"description":["The genes that the Allele is located on"],"id":["allele_of_genes"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"WB:WBGene00001177\""]},"fmin":26380,"seqId":"V","uniqueID":"NC_003283.11:g.26381G>A","alleles":{"meta":{"number":[-1],"description":["The alleles of the variant"],"id":["alleles"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":["\"WB:WBVar00604105\""]}},{"symbol":{"meta":{"number":[1],"description":["The human readable name of the allele"],"id":["symbol"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"n488\""]},"alternative_alleles":{"meta":{"description":"VCF ALT field, list of alternate non-reference alleles called on at least one of the samples"},"values":"T"},"geneLevelConsequence":{"meta":{"number":[-1],"description":["VEP consequence of the variant"],"id":["geneLevelConsequence"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":[["\"splice_acceptor_variant","splice_donor_variant","coding_sequence_variant","intron_variant\""]]},"hgvs_nomenclature":{"meta":{"number":[1],"description":["the HGVS name of the allele"],"id":["hgvs_nomenclature"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"NC_003283.11:g.28912_30730del\""]},"impact":{"meta":{"number":[1],"description":["Variant impact scale"],"id":["impact"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"HIGH\""]},"description":"deletion TAACTTAATAAGTTATGAAAGCAAGAATAAATAGAATGACAGGTTGCTTGAAACGATGGAATGCGATATTTTCTCTTATAGGCTTCTTCTTTTTAGGATTTTTTATAAAAAATGAGGTGTCGGCTGCTACCTGTGAGGTTTTGGCTGCCACCTATAAAGTGTCGGCTGCTACCCATGAGGTGTCCGCTGCTACCCATGAGGTGCCGGCTGCATAGTCTCACAGGTTTAGTGGTTTTCTTGCACGAAATATACTATTTTTGAAAGAAACAGTAAACCCATTCGTGCCTTCATCGTTTCTCAAACTGTCATTTTATGGGTCTTGCCACGAAACTTCTCCTCTCTATCACTCTACGTATCTTTATCTGTCTATAATATTTGGCCGCCCTAGTCTCCCCCTTCCGTCACTGTCTCTCTATGTATTTCAGCAGGCGATCAGCGTGCTCATCCGGAAGTCGAGCAACCCGTGAGCTCATCCTCTCCAGCCACACCGTCCATCAGCGGACCCCCGCCGTGTGCCACGTCATCAGGTTCAACGTCATCAATCACTATTACCACAACGGGATGCTCCACGTCATCGTCGGGACCCTCAAAACATATTCTAGGGGGTAAGAATTAGTTGTTTGCCCTGCGATCCTGAATCAAAATAATTGATATTTCTCTACCCCCGGTAGCTTCTTTTTTTTCGTACTTCCCTCTTCCCCTAACCAACAGAAACGTGCGCTTTTTTCTCCTTTCCTTCTATTTTTCTTTCCATTTCTTTTCTCTTTGGCTTGCCGCGGGCTGCAGCTTTTTTTGAACCACGATTTTTGAATTGATTTATTTTTAATTTTTTTTGGTTTGCAAATTTGGCTAGAAAATATGAAAATTGGAAAATAGTTTTAGAGGAATTTTGTTCAAAAATGTAGGTTAAGATACATTGATACTAAATCCAGGGAAGATATCTAGCGAACGAATCTCATTAGGTGTCGGCTGCTACATGTCGGCTGCACCCCAAAATGTGACGGTTATATACTGTCGGCTGCAACCCATAAGGTGTCGGCTGCAACCCACAAGGTGTCGGCTGCAACCTACAAGGTGTCGGCTGCAACCCACGAATTTTCGGCAGCAACCCACGGGGTGTCGGCTGCCACCCACAAGTTGTCGGCTGTAACCCACAAGGTGTCAGCTGCAACCTACGAAGTGTCGGCTGCTACTTGTCGGTTGCAACCCTGCCCTCACAAAAGCAGCCGGCGTCCAGCGGTTCCCTTATGATTTCTTCAAAAAAAACCGTAAAACTTTCAATTTTTTAGGTGAAATGCCCGCCAAGGAAAACGACGAAGCACATCCGGAGCTAAAGCAGAATTTCATTGCTAAAAATTTGAAAGGATTCGGGTTTTCGAAAAAACAAGTGAGTTTTTTTGAAATTTTCATTTTTTTTCTTCTTTTTCAGATTTCATTATTTCAGGAAACAAACAAAATTTCAAATTTTAACATAGTTTTTAAGGGTTTCCTCTTCCTCCTTATTCTTATTCTTCTCTCTCAAATCCCTTAACTTTTTTATGCATTTGCGCTCTACTGATAATCGAAATTCTCATAAGAGCGTATTTGCACTAAACTTGCCTGCGTTTGAAGTGTTCCTCTTCCTAAGCTGTCTCATAATTTCTGTTCTCTGTTGATTATCATTTTCTTGTTTTTCTGTTTCGTTTCTCTGATTTTCTCGTAGAAAACTGTTGTTTTTTTTCTGATTTCCATGTACTCTTACAACGTACTACTAACTCTCACTTGTTGGGTAGTAGTGCAGATTAAACAAAAAATAAAATCTGAAAGTTTGTAATTTCAGA > T","type":"deletion","symbol_text":{"meta":{"number":[1],"description":["Another human readable representation of the allele"],"id":["symbol_text"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"n488\""]},"fmax":30730,"reference_allele":"TAACTTAATAAGTTATGAAAGCAAGAATAAATAGAATGACAGGTTGCTTGAAACGATGGAATGCGATATTTTCTCTTATAGGCTTCTTCTTTTTAGGATTTTTTATAAAAAATGAGGTGTCGGCTGCTACCTGTGAGGTTTTGGCTGCCACCTATAAAGTGTCGGCTGCTACCCATGAGGTGTCCGCTGCTACCCATGAGGTGCCGGCTGCATAGTCTCACAGGTTTAGTGGTTTTCTTGCACGAAATATACTATTTTTGAAAGAAACAGTAAACCCATTCGTGCCTTCATCGTTTCTCAAACTGTCATTTTATGGGTCTTGCCACGAAACTTCTCCTCTCTATCACTCTACGTATCTTTATCTGTCTATAATATTTGGCCGCCCTAGTCTCCCCCTTCCGTCACTGTCTCTCTATGTATTTCAGCAGGCGATCAGCGTGCTCATCCGGAAGTCGAGCAACCCGTGAGCTCATCCTCTCCAGCCACACCGTCCATCAGCGGACCCCCGCCGTGTGCCACGTCATCAGGTTCAACGTCATCAATCACTATTACCACAACGGGATGCTCCACGTCATCGTCGGGACCCTCAAAACATATTCTAGGGGGTAAGAATTAGTTGTTTGCCCTGCGATCCTGAATCAAAATAATTGATATTTCTCTACCCCCGGTAGCTTCTTTTTTTTCGTACTTCCCTCTTCCCCTAACCAACAGAAACGTGCGCTTTTTTCTCCTTTCCTTCTATTTTTCTTTCCATTTCTTTTCTCTTTGGCTTGCCGCGGGCTGCAGCTTTTTTTGAACCACGATTTTTGAATTGATTTATTTTTAATTTTTTTTGGTTTGCAAATTTGGCTAGAAAATATGAAAATTGGAAAATAGTTTTAGAGGAATTTTGTTCAAAAATGTAGGTTAAGATACATTGATACTAAATCCAGGGAAGATATCTAGCGAACGAATCTCATTAGGTGTCGGCTGCTACATGTCGGCTGCACCCCAAAATGTGACGGTTATATACTGTCGGCTGCAACCCATAAGGTGTCGGCTGCAACCCACAAGGTGTCGGCTGCAACCTACAAGGTGTCGGCTGCAACCCACGAATTTTCGGCAGCAACCCACGGGGTGTCGGCTGCCACCCACAAGTTGTCGGCTGTAACCCACAAGGTGTCAGCTGCAACCTACGAAGTGTCGGCTGCTACTTGTCGGTTGCAACCCTGCCCTCACAAAAGCAGCCGGCGTCCAGCGGTTCCCTTATGATTTCTTCAAAAAAAACCGTAAAACTTTCAATTTTTTAGGTGAAATGCCCGCCAAGGAAAACGACGAAGCACATCCGGAGCTAAAGCAGAATTTCATTGCTAAAAATTTGAAAGGATTCGGGTTTTCGAAAAAACAAGTGAGTTTTTTTGAAATTTTCATTTTTTTTCTTCTTTTTCAGATTTCATTATTTCAGGAAACAAACAAAATTTCAAATTTTAACATAGTTTTTAAGGGTTTCCTCTTCCTCCTTATTCTTATTCTTCTCTCTCAAATCCCTTAACTTTTTTATGCATTTGCGCTCTACTGATAATCGAAATTCTCATAAGAGCGTATTTGCACTAAACTTGCCTGCGTTTGAAGTGTTCCTCTTCCTAAGCTGTCTCATAATTTCTGTTCTCTGTTGATTATCATTTTCTTGTTTTTCTGTTTCGTTTCTCTGATTTTCTCGTAGAAAACTGTTGTTTTTTTTCTGATTTCCATGTACTCTTACAACGTACTACTAACTCTCACTTGTTGGGTAGTAGTGCAGATTAAACAAAAAATAAAATCTGAAAGTTTGTAATTTCAGA","name":"NC_003283.11:g.28912_30730del","allele_of_genes":{"meta":{"number":[1],"description":["The genes that the Allele is located on"],"id":["allele_of_genes"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"WB:WBGene00001177\""]},"fmin":28910,"seqId":"V","uniqueID":"NC_003283.11:g.28912_30730del","alleles":{"meta":{"number":[-1],"description":["The alleles of the variant"],"id":["alleles"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":["\"WB:WBVar00089535\""]}},{"symbol":{"meta":{"number":[1],"description":["The human readable name of the allele"],"id":["symbol"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"md1971\""]},"alternative_alleles":{"meta":{"description":"VCF ALT field, list of alternate non-reference alleles called on at least one of the samples"},"values":"T"},"geneLevelConsequence":{"meta":{"number":[-1],"description":["VEP consequence of the variant"],"id":["geneLevelConsequence"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":["\"stop_gained\""]},"hgvs_nomenclature":{"meta":{"number":[1],"description":["the HGVS name of the allele"],"id":["hgvs_nomenclature"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"NC_003283.11:g.36741C>T\""]},"impact":{"meta":{"number":[1],"description":["Variant impact scale"],"id":["impact"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"HIGH\""]},"description":"SNV C > T","type":"SNV","symbol_text":{"meta":{"number":[1],"description":["Another human readable representation of the allele"],"id":["symbol_text"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"md1971\""]},"fmax":36741,"reference_allele":"C","name":"NC_003283.11:g.36741C>T","allele_of_genes":{"meta":{"number":[1],"description":["The genes that the Allele is located on"],"id":["allele_of_genes"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"WB:WBGene00001177\""]},"fmin":36740,"seqId":"V","uniqueID":"NC_003283.11:g.36741C>T","alleles":{"meta":{"number":[-1],"description":["The alleles of the variant"],"id":["alleles"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":["\"WB:WBVar00088858\""]}},{"symbol":{"meta":{"number":[1],"description":["The human readable name of the allele"],"id":["symbol"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"tm7776\""]},"alternative_alleles":{"meta":{"description":"VCF ALT field, list of alternate non-reference alleles called on at least one of the samples"},"values":"A"},"geneLevelConsequence":{"meta":{"number":[-1],"description":["VEP consequence of the variant"],"id":["geneLevelConsequence"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":[["\"splice_donor_variant","coding_sequence_variant","5_prime_UTR_variant","intron_variant\""]]},"hgvs_nomenclature":{"meta":{"number":[1],"description":["the HGVS name of the allele"],"id":["hgvs_nomenclature"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"NC_003283.11:g.43977_45507del\""]},"impact":{"meta":{"number":[1],"description":["Variant impact scale"],"id":["impact"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"HIGH\""]},"description":"deletion ATATTTAAGGCTTAGGCTTTGGCTCAGGTTTAGGTTTACAAGGCGTTATATTGTTAGGCGTTATCCAACTACTAAATTTATAGTAAGTTTTTTTTTATAGTACTAAATTTATGAACCATATTTCCTGTCAAAATTTTGATTTCGAAATGAAGTATAATACACAAACGGTATATCCTTATCAGTGCTGATCTTTTGAGTTTTCTCCCATAAAATTGTTCAGTTATAAGCTTAATTTCATTTTTACAAGCTGTGCAGTTCTTATCAACCTACGTACAACTCGCTTACCCGCTTATTTTATGACGAAGATCTACGTACAAAAATTGTGAATTTTTGATTTTTGCACACATCAAGAACATATGACATTTCGGTTGAACAAAAAATAAATTTTGAAAATATACATTTTTTGAGGAGTAAAATCGATTTGGGATGACTTTGATTATAGAAAATTGACCAACCAATTATATTTCTGAGGAACATTTTTATTATGTTATAAATTCTCAATTTTTTTTTAATTGACGAAAATGATTTGTAACATTGTTTTATGAAATTTATTTCGCGGGAAAAGCCTGAAAACTTGGCAAACGATTTTTCGAATTTTAAAACAAAAAATAATTTTTTTCAATTCAAAAAATTCAATCTGAAATGTCAAACCTAGAGTATTTATCTGCGTGGGAATTAAAAAAAATTATAACAAAATTTCAAAAGTTCTTTTTTTAACGTTTAGATTTGGTATATTGAGACTGTCTTCTGCCATTTAAGAAAAAAAAAACTCGCAGACGTTTTTCCATCTTAAATTTTTTGTCAAAATTCCAGTAAAAATTCCACATTGTTTCCAGTTTAAAGGGGGTGTAGTTAAACTTTTTATTGTTTTATTTCACTCACAATTGTCTGAAAACACCAAATTTCATAATGAAACTTTTTGAAAACTTCCCAATGGACAAAAAGATTGCGTAAAATATTTTAGAAAATCTATTAGTTCATCAACTTTAAAATACCTAAATCGAGTTGAAAACGCAAGCTAAAATGTTTCCAGCAGCTGCAACATTGGCAAGTCGGTCAAATGTTAAAGTATAACTAATTATAGGTTATTTTTGAGCCGTTATAACTTTTCTTTTAAGAAGTTTTCAAGAAGTTTCATCATGAAATTCAATGTTTTCAGACAATTTTGAGTATAATAAAACAATTAAAAAAAAACTACACCACCTTTAAAGCAAAACCTTGACTTTTTGAAACTAAAAAGTTCAATTTTTTATATTATTAATTCATTTTCATTTTTTTTTTCTTAAACTTTTTGAAAGTCATAAAAATCTCCACCATTTACTACACATTTTATCATACATAATCATGAATTAAAAATTTCGAAATTTCTTTATGGAAAAAGTGAGTCAAACTCGATAGAAACCCCTATGGAGCCAAAGATAACGCCCTTAATTTTCCTTTAAAATGAGGCTCTTGGAACTGATAGAGTAGTTGACGGAAGAAGACGCAATAATGACAACAACTGTTGTTGTTTTAGGACCTACGGTGGGTTA > A","type":"deletion","symbol_text":{"meta":{"number":[1],"description":["Another human readable representation of the allele"],"id":["symbol_text"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"tm7776\""]},"fmax":45507,"reference_allele":"ATATTTAAGGCTTAGGCTTTGGCTCAGGTTTAGGTTTACAAGGCGTTATATTGTTAGGCGTTATCCAACTACTAAATTTATAGTAAGTTTTTTTTTATAGTACTAAATTTATGAACCATATTTCCTGTCAAAATTTTGATTTCGAAATGAAGTATAATACACAAACGGTATATCCTTATCAGTGCTGATCTTTTGAGTTTTCTCCCATAAAATTGTTCAGTTATAAGCTTAATTTCATTTTTACAAGCTGTGCAGTTCTTATCAACCTACGTACAACTCGCTTACCCGCTTATTTTATGACGAAGATCTACGTACAAAAATTGTGAATTTTTGATTTTTGCACACATCAAGAACATATGACATTTCGGTTGAACAAAAAATAAATTTTGAAAATATACATTTTTTGAGGAGTAAAATCGATTTGGGATGACTTTGATTATAGAAAATTGACCAACCAATTATATTTCTGAGGAACATTTTTATTATGTTATAAATTCTCAATTTTTTTTTAATTGACGAAAATGATTTGTAACATTGTTTTATGAAATTTATTTCGCGGGAAAAGCCTGAAAACTTGGCAAACGATTTTTCGAATTTTAAAACAAAAAATAATTTTTTTCAATTCAAAAAATTCAATCTGAAATGTCAAACCTAGAGTATTTATCTGCGTGGGAATTAAAAAAAATTATAACAAAATTTCAAAAGTTCTTTTTTTAACGTTTAGATTTGGTATATTGAGACTGTCTTCTGCCATTTAAGAAAAAAAAAACTCGCAGACGTTTTTCCATCTTAAATTTTTTGTCAAAATTCCAGTAAAAATTCCACATTGTTTCCAGTTTAAAGGGGGTGTAGTTAAACTTTTTATTGTTTTATTTCACTCACAATTGTCTGAAAACACCAAATTTCATAATGAAACTTTTTGAAAACTTCCCAATGGACAAAAAGATTGCGTAAAATATTTTAGAAAATCTATTAGTTCATCAACTTTAAAATACCTAAATCGAGTTGAAAACGCAAGCTAAAATGTTTCCAGCAGCTGCAACATTGGCAAGTCGGTCAAATGTTAAAGTATAACTAATTATAGGTTATTTTTGAGCCGTTATAACTTTTCTTTTAAGAAGTTTTCAAGAAGTTTCATCATGAAATTCAATGTTTTCAGACAATTTTGAGTATAATAAAACAATTAAAAAAAAACTACACCACCTTTAAAGCAAAACCTTGACTTTTTGAAACTAAAAAGTTCAATTTTTTATATTATTAATTCATTTTCATTTTTTTTTTCTTAAACTTTTTGAAAGTCATAAAAATCTCCACCATTTACTACACATTTTATCATACATAATCATGAATTAAAAATTTCGAAATTTCTTTATGGAAAAAGTGAGTCAAACTCGATAGAAACCCCTATGGAGCCAAAGATAACGCCCTTAATTTTCCTTTAAAATGAGGCTCTTGGAACTGATAGAGTAGTTGACGGAAGAAGACGCAATAATGACAACAACTGTTGTTGTTTTAGGACCTACGGTGGGTTA","name":"NC_003283.11:g.43977_45507del","allele_of_genes":{"meta":{"number":[1],"description":["The genes that the Allele is located on"],"id":["allele_of_genes"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"WB:WBGene00015152\""]},"fmin":43975,"seqId":"V","uniqueID":"NC_003283.11:g.43977_45507del","alleles":{"meta":{"number":[-1],"description":["The alleles of the variant"],"id":["alleles"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":["\"WB:WBVar02144957\""]}}];
       // return [{"symbol":{"meta":{"number":[1],"description":["The human readable name of the allele"],"id":["symbol"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"tm6709\""]},"alternative_alleles":{"meta":{"description":"VCF ALT field, list of alternate non-reference alleles called on at least one of the samples"},"values":"A"},"geneLevelConsequence":{"meta":{"number":[-1],"description":["VEP consequence of the variant"],"id":["geneLevelConsequence"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":[["\"splice_acceptor_variant","splice_donor_variant","stop_lost","inframe_deletion","intron_variant\""]]},"hgvs_nomenclature":{"meta":{"number":[1],"description":["the HGVS name of the allele"],"id":["hgvs_nomenclature"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"NC_003283.11:g.7355_7776del\""]},"impact":{"meta":{"number":[1],"description":["Variant impact scale"],"id":["impact"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"HIGH\""]},"description":"deletion ACCAGTAGTGATCGAGCAATTGCGTTCTTCTCTGAAGCTTCTATGATTCATTGAATAAAATATATTTCTCAAAACGTACTTGCTTATCGACAACAACCAACCAACGTCCACCTTGAACGTTGTTGACGTCCTCCCACATTGGCTTGATTCCTTCCTTGAACAAGTAATAATCGGATCCCCAGTTCAATCCTCCGGCAGACTGAATGTGATTGTACAGCGACCAGAAGTCCTCGACAGTGTCGAAAAGTGAAACCATCTGGAAAAAATCGATAAAAGACGTATTTAAAAATCTTCTACCTTCAGACAATCCTCCCATTCCTTGTTACGGTCAGCTTTCAAGTACCAGAGAGCCCAGCGATTCTGGAGGGGGTGTCTGGTGAGAAGCTCTGGAGGAACTGAAGCATCGGACGCATTCACATCGCC > A","type":"deletion","symbol_text":{"meta":{"number":[1],"description":["Another human readable representation of the allele"],"id":["symbol_text"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"tm6709\""]},"fmax":7776,"reference_allele":"ACCAGTAGTGATCGAGCAATTGCGTTCTTCTCTGAAGCTTCTATGATTCATTGAATAAAATATATTTCTCAAAACGTACTTGCTTATCGACAACAACCAACCAACGTCCACCTTGAACGTTGTTGACGTCCTCCCACATTGGCTTGATTCCTTCCTTGAACAAGTAATAATCGGATCCCCAGTTCAATCCTCCGGCAGACTGAATGTGATTGTACAGCGACCAGAAGTCCTCGACAGTGTCGAAAAGTGAAACCATCTGGAAAAAATCGATAAAAGACGTATTTAAAAATCTTCTACCTTCAGACAATCCTCCCATTCCTTGTTACGGTCAGCTTTCAAGTACCAGAGAGCCCAGCGATTCTGGAGGGGGTGTCTGGTGAGAAGCTCTGGAGGAACTGAAGCATCGGACGCATTCACATCGCC","name":"NC_003283.11:g.7355_7776del","allele_of_genes":{"meta":{"number":[1],"description":["The genes that the Allele is located on"],"id":["allele_of_genes"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"WB:WBGene00002061\""]},"fmin":7353,"seqId":"V","uniqueID":"NC_003283.11:g.7355_7776del","alleles":{"meta":{"number":[-1],"description":["The alleles of the variant"],"id":["alleles"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":["\"WB:WBVar02125540\""]}},{"symbol":{"meta":{"number":[1],"description":["The human readable name of the allele"],"id":["symbol"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"tm774\""]},"alternative_alleles":{"meta":{"description":"VCF ALT field, list of alternate non-reference alleles called on at least one of the samples"},"values":"C"},"geneLevelConsequence":{"meta":{"number":[-1],"description":["VEP consequence of the variant"],"id":["geneLevelConsequence"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":["\"intron_variant\""]},"hgvs_nomenclature":{"meta":{"number":[1],"description":["the HGVS name of the allele"],"id":["hgvs_nomenclature"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"NC_003283.11:g.23999_24591del\""]},"impact":{"meta":{"number":[1],"description":["Variant impact scale"],"id":["impact"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"MODIFIER\""]},"description":"deletion CTAACAGAAAATGTACTATTACACCGATTATTAGTACTTATAATTTTTTTTTTGAAAGTTGGAAAATATTTTAAAGTTTGTAAAAATTCTATAAATAATATTGAATATTATAACGGTAAAATATAATATTTATAGAATTTTTTCAATTTTCAAAAAAGTACTGTTTGTCATTTGTTTTTGTTTGTTAGAATACCTTTTAATAGCGGTACCGAAATCTGGGAAATATTTTCAAATGACTCCAAATTTTCGCCTGATTCCGAATATCTATGTGAAAAAAAGTGAAAAAAAACATCCCTTATTTTATATTTCATTTCCAATTCCCAATTTCATTTGTGCCCGCATTACTTTTTTCAAACGCGCGCCCAAATAAATTTTCATTAGAGCGCGTTTGCTTTGTGTCGATTTACGGGAGCTCTCCATTTACAAAAAAAATTAACGCTTTTTTTAAACAGTTTTTAAGCGAGTTTCCTACTTTTTTAATCGATTTCAGCGGAACGGTTTTTCTTGAAATATCGTTTTTAAGGTCAATTGATATGATAATAAAATATATTGTTGACCTTAAAACACATTAAAAAAACTTCGACAAGAACCGAC > C","type":"deletion","symbol_text":{"meta":{"number":[1],"description":["Another human readable representation of the allele"],"id":["symbol_text"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"tm774\""]},"fmax":24591,"reference_allele":"CTAACAGAAAATGTACTATTACACCGATTATTAGTACTTATAATTTTTTTTTTGAAAGTTGGAAAATATTTTAAAGTTTGTAAAAATTCTATAAATAATATTGAATATTATAACGGTAAAATATAATATTTATAGAATTTTTTCAATTTTCAAAAAAGTACTGTTTGTCATTTGTTTTTGTTTGTTAGAATACCTTTTAATAGCGGTACCGAAATCTGGGAAATATTTTCAAATGACTCCAAATTTTCGCCTGATTCCGAATATCTATGTGAAAAAAAGTGAAAAAAAACATCCCTTATTTTATATTTCATTTCCAATTCCCAATTTCATTTGTGCCCGCATTACTTTTTTCAAACGCGCGCCCAAATAAATTTTCATTAGAGCGCGTTTGCTTTGTGTCGATTTACGGGAGCTCTCCATTTACAAAAAAAATTAACGCTTTTTTTAAACAGTTTTTAAGCGAGTTTCCTACTTTTTTAATCGATTTCAGCGGAACGGTTTTTCTTGAAATATCGTTTTTAAGGTCAATTGATATGATAATAAAATATATTGTTGACCTTAAAACACATTAAAAAAACTTCGACAAGAACCGAC","name":"NC_003283.11:g.23999_24591del","allele_of_genes":{"meta":{"number":[1],"description":["The genes that the Allele is located on"],"id":["allele_of_genes"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"WB:WBGene00001177\""]},"fmin":23997,"seqId":"V","uniqueID":"NC_003283.11:g.23999_24591del","alleles":{"meta":{"number":[-1],"description":["The alleles of the variant"],"id":["alleles"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":["\"WB:WBVar00249802\""]}},{"symbol":{"meta":{"number":[1],"description":["The human readable name of the allele"],"id":["symbol"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"nj77\""]},"alternative_alleles":{"meta":{"description":"VCF ALT field, list of alternate non-reference alleles called on at least one of the samples"},"values":"A"},"geneLevelConsequence":{"meta":{"number":[-1],"description":["VEP consequence of the variant"],"id":["geneLevelConsequence"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":["\"splice_donor_variant\""]},"hgvs_nomenclature":{"meta":{"number":[1],"description":["the HGVS name of the allele"],"id":["hgvs_nomenclature"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"NC_003283.11:g.26381G>A\""]},"impact":{"meta":{"number":[1],"description":["Variant impact scale"],"id":["impact"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"HIGH\""]},"description":"SNV G > A","type":"SNV","symbol_text":{"meta":{"number":[1],"description":["Another human readable representation of the allele"],"id":["symbol_text"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"nj77\""]},"fmax":26381,"reference_allele":"G","name":"NC_003283.11:g.26381G>A","allele_of_genes":{"meta":{"number":[1],"description":["The genes that the Allele is located on"],"id":["allele_of_genes"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"WB:WBGene00001177\""]},"fmin":26380,"seqId":"V","uniqueID":"NC_003283.11:g.26381G>A","alleles":{"meta":{"number":[-1],"description":["The alleles of the variant"],"id":["alleles"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":["\"WB:WBVar00604105\""]}},{"symbol":{"meta":{"number":[1],"description":["The human readable name of the allele"],"id":["symbol"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"n488\""]},"alternative_alleles":{"meta":{"description":"VCF ALT field, list of alternate non-reference alleles called on at least one of the samples"},"values":"T"},"geneLevelConsequence":{"meta":{"number":[-1],"description":["VEP consequence of the variant"],"id":["geneLevelConsequence"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":[["\"splice_acceptor_variant","splice_donor_variant","coding_sequence_variant","intron_variant\""]]},"hgvs_nomenclature":{"meta":{"number":[1],"description":["the HGVS name of the allele"],"id":["hgvs_nomenclature"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"NC_003283.11:g.28912_30730del\""]},"impact":{"meta":{"number":[1],"description":["Variant impact scale"],"id":["impact"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"HIGH\""]},"description":"deletion TAACTTAATAAGTTATGAAAGCAAGAATAAATAGAATGACAGGTTGCTTGAAACGATGGAATGCGATATTTTCTCTTATAGGCTTCTTCTTTTTAGGATTTTTTATAAAAAATGAGGTGTCGGCTGCTACCTGTGAGGTTTTGGCTGCCACCTATAAAGTGTCGGCTGCTACCCATGAGGTGTCCGCTGCTACCCATGAGGTGCCGGCTGCATAGTCTCACAGGTTTAGTGGTTTTCTTGCACGAAATATACTATTTTTGAAAGAAACAGTAAACCCATTCGTGCCTTCATCGTTTCTCAAACTGTCATTTTATGGGTCTTGCCACGAAACTTCTCCTCTCTATCACTCTACGTATCTTTATCTGTCTATAATATTTGGCCGCCCTAGTCTCCCCCTTCCGTCACTGTCTCTCTATGTATTTCAGCAGGCGATCAGCGTGCTCATCCGGAAGTCGAGCAACCCGTGAGCTCATCCTCTCCAGCCACACCGTCCATCAGCGGACCCCCGCCGTGTGCCACGTCATCAGGTTCAACGTCATCAATCACTATTACCACAACGGGATGCTCCACGTCATCGTCGGGACCCTCAAAACATATTCTAGGGGGTAAGAATTAGTTGTTTGCCCTGCGATCCTGAATCAAAATAATTGATATTTCTCTACCCCCGGTAGCTTCTTTTTTTTCGTACTTCCCTCTTCCCCTAACCAACAGAAACGTGCGCTTTTTTCTCCTTTCCTTCTATTTTTCTTTCCATTTCTTTTCTCTTTGGCTTGCCGCGGGCTGCAGCTTTTTTTGAACCACGATTTTTGAATTGATTTATTTTTAATTTTTTTTGGTTTGCAAATTTGGCTAGAAAATATGAAAATTGGAAAATAGTTTTAGAGGAATTTTGTTCAAAAATGTAGGTTAAGATACATTGATACTAAATCCAGGGAAGATATCTAGCGAACGAATCTCATTAGGTGTCGGCTGCTACATGTCGGCTGCACCCCAAAATGTGACGGTTATATACTGTCGGCTGCAACCCATAAGGTGTCGGCTGCAACCCACAAGGTGTCGGCTGCAACCTACAAGGTGTCGGCTGCAACCCACGAATTTTCGGCAGCAACCCACGGGGTGTCGGCTGCCACCCACAAGTTGTCGGCTGTAACCCACAAGGTGTCAGCTGCAACCTACGAAGTGTCGGCTGCTACTTGTCGGTTGCAACCCTGCCCTCACAAAAGCAGCCGGCGTCCAGCGGTTCCCTTATGATTTCTTCAAAAAAAACCGTAAAACTTTCAATTTTTTAGGTGAAATGCCCGCCAAGGAAAACGACGAAGCACATCCGGAGCTAAAGCAGAATTTCATTGCTAAAAATTTGAAAGGATTCGGGTTTTCGAAAAAACAAGTGAGTTTTTTTGAAATTTTCATTTTTTTTCTTCTTTTTCAGATTTCATTATTTCAGGAAACAAACAAAATTTCAAATTTTAACATAGTTTTTAAGGGTTTCCTCTTCCTCCTTATTCTTATTCTTCTCTCTCAAATCCCTTAACTTTTTTATGCATTTGCGCTCTACTGATAATCGAAATTCTCATAAGAGCGTATTTGCACTAAACTTGCCTGCGTTTGAAGTGTTCCTCTTCCTAAGCTGTCTCATAATTTCTGTTCTCTGTTGATTATCATTTTCTTGTTTTTCTGTTTCGTTTCTCTGATTTTCTCGTAGAAAACTGTTGTTTTTTTTCTGATTTCCATGTACTCTTACAACGTACTACTAACTCTCACTTGTTGGGTAGTAGTGCAGATTAAACAAAAAATAAAATCTGAAAGTTTGTAATTTCAGA > T","type":"deletion","symbol_text":{"meta":{"number":[1],"description":["Another human readable representation of the allele"],"id":["symbol_text"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"n488\""]},"fmax":30730,"reference_allele":"TAACTTAATAAGTTATGAAAGCAAGAATAAATAGAATGACAGGTTGCTTGAAACGATGGAATGCGATATTTTCTCTTATAGGCTTCTTCTTTTTAGGATTTTTTATAAAAAATGAGGTGTCGGCTGCTACCTGTGAGGTTTTGGCTGCCACCTATAAAGTGTCGGCTGCTACCCATGAGGTGTCCGCTGCTACCCATGAGGTGCCGGCTGCATAGTCTCACAGGTTTAGTGGTTTTCTTGCACGAAATATACTATTTTTGAAAGAAACAGTAAACCCATTCGTGCCTTCATCGTTTCTCAAACTGTCATTTTATGGGTCTTGCCACGAAACTTCTCCTCTCTATCACTCTACGTATCTTTATCTGTCTATAATATTTGGCCGCCCTAGTCTCCCCCTTCCGTCACTGTCTCTCTATGTATTTCAGCAGGCGATCAGCGTGCTCATCCGGAAGTCGAGCAACCCGTGAGCTCATCCTCTCCAGCCACACCGTCCATCAGCGGACCCCCGCCGTGTGCCACGTCATCAGGTTCAACGTCATCAATCACTATTACCACAACGGGATGCTCCACGTCATCGTCGGGACCCTCAAAACATATTCTAGGGGGTAAGAATTAGTTGTTTGCCCTGCGATCCTGAATCAAAATAATTGATATTTCTCTACCCCCGGTAGCTTCTTTTTTTTCGTACTTCCCTCTTCCCCTAACCAACAGAAACGTGCGCTTTTTTCTCCTTTCCTTCTATTTTTCTTTCCATTTCTTTTCTCTTTGGCTTGCCGCGGGCTGCAGCTTTTTTTGAACCACGATTTTTGAATTGATTTATTTTTAATTTTTTTTGGTTTGCAAATTTGGCTAGAAAATATGAAAATTGGAAAATAGTTTTAGAGGAATTTTGTTCAAAAATGTAGGTTAAGATACATTGATACTAAATCCAGGGAAGATATCTAGCGAACGAATCTCATTAGGTGTCGGCTGCTACATGTCGGCTGCACCCCAAAATGTGACGGTTATATACTGTCGGCTGCAACCCATAAGGTGTCGGCTGCAACCCACAAGGTGTCGGCTGCAACCTACAAGGTGTCGGCTGCAACCCACGAATTTTCGGCAGCAACCCACGGGGTGTCGGCTGCCACCCACAAGTTGTCGGCTGTAACCCACAAGGTGTCAGCTGCAACCTACGAAGTGTCGGCTGCTACTTGTCGGTTGCAACCCTGCCCTCACAAAAGCAGCCGGCGTCCAGCGGTTCCCTTATGATTTCTTCAAAAAAAACCGTAAAACTTTCAATTTTTTAGGTGAAATGCCCGCCAAGGAAAACGACGAAGCACATCCGGAGCTAAAGCAGAATTTCATTGCTAAAAATTTGAAAGGATTCGGGTTTTCGAAAAAACAAGTGAGTTTTTTTGAAATTTTCATTTTTTTTCTTCTTTTTCAGATTTCATTATTTCAGGAAACAAACAAAATTTCAAATTTTAACATAGTTTTTAAGGGTTTCCTCTTCCTCCTTATTCTTATTCTTCTCTCTCAAATCCCTTAACTTTTTTATGCATTTGCGCTCTACTGATAATCGAAATTCTCATAAGAGCGTATTTGCACTAAACTTGCCTGCGTTTGAAGTGTTCCTCTTCCTAAGCTGTCTCATAATTTCTGTTCTCTGTTGATTATCATTTTCTTGTTTTTCTGTTTCGTTTCTCTGATTTTCTCGTAGAAAACTGTTGTTTTTTTTCTGATTTCCATGTACTCTTACAACGTACTACTAACTCTCACTTGTTGGGTAGTAGTGCAGATTAAACAAAAAATAAAATCTGAAAGTTTGTAATTTCAGA","name":"NC_003283.11:g.28912_30730del","allele_of_genes":{"meta":{"number":[1],"description":["The genes that the Allele is located on"],"id":["allele_of_genes"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"WB:WBGene00001177\""]},"fmin":28910,"seqId":"V","uniqueID":"NC_003283.11:g.28912_30730del","alleles":{"meta":{"number":[-1],"description":["The alleles of the variant"],"id":["alleles"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":["\"WB:WBVar00089535\""]}}];
@@ -95,4 +95,3 @@ class ApolloService
     }
 }
 
-export { ApolloService }
